test: split cleanup admission inflight cases under 300 lines
Divide in-flight remove, pass-cancel completed omission, and start-event
append hold coverage into separate files so none exceeds the AGENTS.md
subsystem test-file limit.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
